requirements.txt
pandas>=1.5.0,<2.0.0
numpy>=1.21.0,<1.25.0
scikit-learn>=1.0.0,<1.3.0
requests>=2.25.0
flask>=2.0.0
psycopg2-binary>=2.8.6
joblib>=1.0.0
sqlalchemy>=1.4.0,<2.0.0
plotly>=5.0.0 